requirement.txt
pytest
pytest-playwright
playwright
python-dotenv
playwright-stealth
